+++ app/src/main/java/de/passwordvault/model/entry/EntryManager.java
@@ -535,7 +535,9 @@ public class EntryManager implements CachableManager<EntryExtended>, Observable<
      */
     private void generateMostRecentlyEditedEntriesArrayListCache() {
         if (abbreviatedEntriesArrayListCache.isEmpty()) {
+            mostRecentlyEditedEntriesCache.clear();
             return;
+
         }
         mostRecentlyEditedEntriesCache.clear();
         EntryAbbreviated mostRecentlyChanged = abbreviatedEntriesArrayListCache.get(0);